Removing redundant explicit casting

+++ src/CartBlock.C
@@ -111,11 +111,9 @@ void CartBlock::getInterpolatedData(
             }
 
             for (i = 0; i < listptr->nweights; i++) {
-                int const cell_index =
-                    static_cast<int>(cart_utils::get_cell_index(
-                        dims[0], dims[1], nf,
-                        listptr->inode[static_cast<int>(3 * i)],
-                        listptr->inode[3 * i + 1], listptr->inode[3 * i + 2]));
+                int const cell_index = (cart_utils::get_cell_index(
+                    dims[0], dims[1], nf, listptr->inode[3 * i],
+                    listptr->inode[3 * i + 1], listptr->inode[3 * i + 2]));
                 for (n = 0; n < nvar_cell; n++) {
                     weight = listptr->weights[i];
                     qq[n] += qcell[cell_index + ncell_nf * n] * weight;
@@ -171,8 +169,7 @@ void CartBlock::preprocess(CartGrid* cg)
     for (int n = 0; n < 3; n++) {
         dx[n] = cg->dx[3 * global_id + n];
     }
-    dims[0] = cg->ihi[static_cast<int>(3 * global_id)] -
-              cg->ilo[static_cast<int>(3 * global_id)] + 1;
+    dims[0] = cg->ihi[3 * global_id] - cg->ilo[3 * global_id] + 1;
     dims[1] = cg->ihi[3 * global_id + 1] - cg->ilo[3 * global_id + 1] + 1;
     dims[2] = cg->ihi[3 * global_id + 2] - cg->ilo[3 * global_id + 2] + 1;
     nf = cg->nf;
@@ -260,10 +257,10 @@ void CartBlock::insertInInterpList(
     }
     if (donor_frac == nullptr) {
         listptr->nweights = 8;
-        listptr->weights = (double*)malloc(
-            sizeof(double) * (static_cast<int>(listptr->nweights * 2)));
-        listptr->inode = (int*)malloc(
-            sizeof(int) * (static_cast<int>(listptr->nweights * 2 * 3)));
+        listptr->weights =
+            (double*)malloc(sizeof(double) * (listptr->nweights * 2));
+        listptr->inode =
+            (int*)malloc(sizeof(int) * (listptr->nweights * 2 * 3));
 
         cart_interp::linear_interpolation(
             nf, ix, dims, rst, &(listptr->nweights), listptr->inode,
