+++ src/bookKeeping.C
@@ -42,7 +42,7 @@ void MeshBlock::getDonorPacket(PACKET* sndPack, int nsend) const
     //
     for (i = 0; i < nsearch; i++) {
         if (donorId[i] > -1) {
-            k = isearch[static_cast<int>(2 * i)];
+            k = isearch[2 * i];
             sndPack[k].nints += 3;
             sndPack[k].nreals++;
         }
@@ -59,7 +59,7 @@ void MeshBlock::getDonorPacket(PACKET* sndPack, int nsend) const
     //
     for (i = 0; i < nsearch; i++) {
         if (donorId[i] > -1) {
-            k = isearch[static_cast<int>(2 * i)];
+            k = isearch[2 * i];
             sndPack[k].intData[icount[k]++] = meshtag;            // mesh tag
             sndPack[k].intData[icount[k]++] = isearch[2 * i + 1]; // point id
             sndPack[k].intData[icount[k]++] = i; // point id on the donor side
@@ -76,7 +76,7 @@ void MeshBlock::getMBDonorPktSizes(
 {
     for (int i = 0; i < nsearch; i++) {
         if (donorId[i] > -1) {
-            int const ii = isearch[static_cast<int>(3 * i)];
+            int const ii = isearch[3 * i];
             nints[ii] += 4;
             nreals[ii] += 2;
         }
@@ -93,7 +93,7 @@ void MeshBlock::getMBDonorPackets(
             continue;
         }
 
-        int const k = isearch[static_cast<int>(3 * i)];
+        int const k = isearch[3 * i];
         int& ix = ixOffset[k];
         int& rx = rxOffset[k];
 
@@ -180,9 +180,7 @@ void MeshBlock::processDonors(
         if (verbose != 0) TRACEI(iblank[i]);
         if (verbose != 0) TRACED(nodeRes[i]);
         if (verbose != 0) {
-            printf(
-                "%f %f %f\n", x[static_cast<int>(3 * i)], x[3 * i + 1],
-                x[3 * i + 2]);
+            printf("%f %f %f\n", x[3 * i], x[3 * i + 1], x[3 * i + 2]);
         }
         if (donorList[i] == nullptr) {
             if (verbose != 0) {
@@ -191,8 +189,8 @@ void MeshBlock::processDonors(
             for (j = 0; j < nmesh; j++) {
                 if (j != (meshtag - BASE) && (holemap[j].existWall != 0)) {
                     if (checkHoleMap(
-                            &x[static_cast<int>(3 * i)], holemap[j].nx,
-                            holemap[j].sam, holemap[j].extents) != 0) {
+                            &x[3 * i], holemap[j].nx, holemap[j].sam,
+                            holemap[j].extents) != 0) {
                         iblank[i] = 0;
                         break;
                     }
@@ -220,8 +218,8 @@ void MeshBlock::processDonors(
                 if (j != (meshtag - BASE) && (holemap[j].existWall != 0)) {
                     if (iflag[j] == 0) {
                         if (checkHoleMap(
-                                &x[static_cast<int>(3 * i)], holemap[j].nx,
-                                holemap[j].sam, holemap[j].extents) != 0) {
+                                &x[3 * i], holemap[j].nx, holemap[j].sam,
+                                holemap[j].extents) != 0) {
                             iblank[i] = 0;
                             break;
                         }
@@ -243,8 +241,7 @@ void MeshBlock::processDonors(
                     myid, wbcnode[i] - BASE, donorList[wbcnode[i] - BASE]);
                 ii = wbcnode[i] - BASE;
                 printf(
-                    "xloc=%e %e %e\n", x[static_cast<int>(3 * ii)],
-                    x[3 * ii + 1], x[3 * ii + 2]);
+                    "xloc=%e %e %e\n", x[3 * ii], x[3 * ii + 1], x[3 * ii + 2]);
                 printf(
                     "Computations will continue, but may suffer from accuracy "
                     "problems\n");
@@ -408,8 +405,8 @@ void MeshBlock::processDonors(
 
             for (j = 0; j < nmesh; j++) {
                 if (j != (meshtag - BASE) && (holemap[j].existWall != 0U)) {
-                    int const SB_val = checkAdaptiveHoleMap(
-                        &x[static_cast<int>(3 * i)], &holemap[j]);
+                    int const SB_val =
+                        checkAdaptiveHoleMap(&x[3 * i], &holemap[j]);
                     if (SB_val != OUTSIDE_SB) {
                         iblank[i] = 0;
                         break;
@@ -446,8 +443,8 @@ void MeshBlock::processDonors(
                     if (iflag[j] == 0) {
                         // body{j} does NOT have candidate so check if point is
                         // INSIDE SB
-                        int const SB_val = checkAdaptiveHoleMap(
-                            &x[static_cast<int>(3 * i)], &holemap[j]);
+                        int const SB_val =
+                            checkAdaptiveHoleMap(&x[3 * i], &holemap[j]);
                         if (SB_val != OUTSIDE_SB) {
                             iblank[i] = 0;
                             break;
@@ -472,8 +469,8 @@ void MeshBlock::processDonors(
                     myid, wbcnode[i] - BASE, donorList[wbcnode[i] - BASE]);
                 ii = wbcnode[i] - BASE;
                 fprintf(
-                    stderr, "xloc=%e %e %e\n", x[static_cast<int>(3 * ii)],
-                    x[3 * ii + 1], x[3 * ii + 2]);
+                    stderr, "xloc=%e %e %e\n", x[3 * ii], x[3 * ii + 1],
+                    x[3 * ii + 2]);
                 fprintf(
                     stderr,
                     "Computations will continue, but may suffer from accuracy "
@@ -653,7 +650,7 @@ void MeshBlock::findInterpData(int* recid, int irecord, double receptorRes2)
     // if (myid==63 && irecord==3224) verbose=1;
     // if (myid==1 && irecord==158192 && meshtag==1) verbose=1;
     receptorRes = fabs(receptorRes2);
-    procid = isearch[static_cast<int>(3 * irecord)];
+    procid = isearch[3 * irecord];
     pointid = isearch[3 * irecord + 1];
     blockid = isearch[3 * irecord + 2];
     meshtagrecv = tagsearch[irecord];
