+++ src/holeMap.C
@@ -126,9 +126,8 @@ void tioga::getHoleMap()
     MPI_Allreduce(
         bboxLocal, bboxGlobal, 3 * maxtag, MPI_DOUBLE, MPI_MIN, scomm);
     MPI_Allreduce(
-        &(bboxLocal[static_cast<int>(3 * maxtag)]),
-        &(bboxGlobal[static_cast<int>(3 * maxtag)]), 3 * maxtag, MPI_DOUBLE,
-        MPI_MAX, scomm);
+        &(bboxLocal[3 * maxtag]), &(bboxGlobal[3 * maxtag]), 3 * maxtag,
+        MPI_DOUBLE, MPI_MAX, scomm);
     //
     // find the bounding box for each mesh
     // from the globally reduced data
@@ -763,7 +762,7 @@ void tioga::getAdaptiveHoleMap()
                         elvl->octants[j].z = lvl->octants[j].z;
                         elvl->octants[j].filltype = lvl->octants[j].filltype;
                         elvl->octants[j].leafflag =
-                            static_cast<uint8_t>(lvl->octants[j].refined) == 0U;
+                            lvl->octants[j].refined == 0U;
                         if (lvl->octants[j].refined != 0U) {
                             for (c = 0; c < OCTANT_CHILDREN; c++) {
                                 elvl->octants[j].children[c] =
@@ -981,7 +980,7 @@ void writePointsVolume(
     for (pt = 0, k = 0; k < npts1d; ++k) {
         for (j = 0; j < npts1d; ++j) {
             for (i = 0; i < npts1d; ++i) {
-                double* co = &x[static_cast<int>(3 * pt)];
+                double* co = &x[3 * pt];
                 fprintf(
                     fp, "%f %f %f %d %d %d\n", co[0], co[1], co[2], type, level,
                     id);
@@ -1008,7 +1007,7 @@ void writePointsVolumeBrick(
     for (pt = 0, k = 0; k < npts1d; ++k) {
         for (j = 0; j < npts1d; ++j) {
             for (i = 0; i < npts1d; ++i) {
-                double* co = &x[static_cast<int>(3 * pt)];
+                double* co = &x[3 * pt];
                 fprintf(fp, "%f ", co[0]);
                 pt++;
             }
@@ -1020,7 +1019,7 @@ void writePointsVolumeBrick(
     for (pt = 0, k = 0; k < npts1d; ++k) {
         for (j = 0; j < npts1d; ++j) {
             for (i = 0; i < npts1d; ++i) {
-                double* co = &x[static_cast<int>(3 * pt)];
+                double* co = &x[3 * pt];
                 fprintf(fp, "%f ", co[1]);
                 pt++;
             }
@@ -1032,7 +1031,7 @@ void writePointsVolumeBrick(
     for (pt = 0, k = 0; k < npts1d; ++k) {
         for (j = 0; j < npts1d; ++j) {
             for (i = 0; i < npts1d; ++i) {
-                double* co = &x[static_cast<int>(3 * pt)];
+                double* co = &x[3 * pt];
                 fprintf(fp, "%f ", co[2]);
                 pt++;
             }
